renaming `prototype` to `model`
